created the searches firectory and moved EqSearch.java file inside of it
